preparation for virtual logs functions
